refactor: improved file structure
